requirements.txt
torch>=2.0.0
numpy
matplotlib
scikit-learn
wandb
albumentations
Pillow
gdown